@@ -1,9 +1,9 @@
 source 'https://rubygems.org'
 
-gem 'rails', '~> 5.2'
+gem 'rails', '~> 6.1'
 
 gem 'activerecord-session_store'
-gem 'bcrypt', '~> 3.1.7'
+gem 'bcrypt'
 gem 'cancan'
 gem 'coffee-rails'
 gem 'devise'
@@ -12,25 +12,31 @@ gem 'exception_notification'
 gem 'git'
 gem 'haml'
 gem 'iso_country_codes'
-gem 'jbuilder', '~> 2.0'
+gem 'jbuilder'
 gem 'paranoia'
 gem 'pg' #, '~> 0.2'
-gem 'rails-i18n', '~> 5.1.1'
-gem 'settingslogic'
-gem 'uglifier', '>= 1.3.0'
+gem 'rails-i18n'
+gem 'uglifier'
 gem 'will_paginate'
 gem 'will_paginate-bootstrap'
-gem 'sdoc', group: :doc
 
-gem 'sprockets', '~> 3'
+# gem 'sdoc', group: :doc
+gem 'ffi', '1.16.3'
+gem 'bootsnap'
+gem 'psych', '~> 3'
+
+
+
+
+gem 'sprockets'
 gem 'rack-attack'
 
 gem 'turbolinks'
 gem 'jquery-rails'
 gem 'jquery-ui-rails'
-gem 'jquery-minicolors-rails', '~> 2.2'
+gem 'jquery-minicolors-rails'
 gem 'jquery-turbolinks'
-gem 'foundation-rails', '= 5.5.1.2'
+gem 'foundation-rails'
 gem 'foundation-icons-sass-rails'
 gem 'foundation-datetimepicker-rails'
 gem 'font-awesome-rails'
@@ -43,9 +49,10 @@ gem "mini_magick"
 
 gem 'datagrid'
 
-gem 'coveralls_reborn', '~> 0.12.0', require: false
+gem 'coveralls_reborn', require: false
 
 group :development do
+  gem 'listen'
   gem 'rails_layout'
   gem 'haml-rails'
   gem 'bootstrap-generators'
@@ -64,7 +71,7 @@ group :test, :development do
   gem 'rspec-rails'
   gem 'rails-controller-testing'
   # gem 'capybara-webkit'
-  gem 'webdrivers', '~> 4.0', require: false
+  gem 'webdrivers'
   gem 'launchy'
   gem 'unicorn'
   gem 'unicorn-rails'
