@@ -1,4 +1,4 @@
-class CreateRoles < ActiveRecord::Migration
+class CreateRoles < ActiveRecord::Migration[4.2]
   def self.up
     create_table :roles, :force => true do |t|
       t.string   :name,              :limit => 40
