@@ -1,4 +1,4 @@
-class AddPTypeToPost < ActiveRecord::Migration
+class AddPTypeToPost < ActiveRecord::Migration[4.2]
   def change
     add_column :posts, :ptype, :string
   end
