@@ -1,4 +1,4 @@
-class AddStatusToPost < ActiveRecord::Migration
+class AddStatusToPost < ActiveRecord::Migration[4.2]
   def change
     add_column :posts, :status, :boolean
   end
